+++ lib/bundler/definition.rb
@@ -21,7 +21,7 @@ module Bundler
       if check
         hash = Digest::SHA1.hexdigest(File.read("#{Bundler.root}/Gemfile"))
         unless locked_definition.hash == hash
-          raise GemfileError, "You changed your Gemfile after locking. Please relock using `bundle lock`"
+          raise GemfileChanged, "You changed your Gemfile after locking. Please relock using `bundle lock`"
         end
       end