+++ lib/bundler/shared_helpers.rb
@@ -1,4 +1,6 @@
 require 'pathname'
+require 'rubygems'
+Gem.source_index # ensure Rubygems is fully loaded in Ruby 1.9
 
 module Gem
   class Dependency
@@ -11,6 +13,28 @@ module Gem
 end
 
 module Bundler
+  class Specification < Gem::Specification
+    attr_accessor :relative_loaded_from
+
+    def self.from_gemspec(gemspec)
+      spec = allocate
+      gemspec.instance_variables.each do |ivar|
+        spec.instance_variable_set(ivar, gemspec.instance_variable_get(ivar))
+      end
+      spec
+    end
+
+    def loaded_from
+      return super unless relative_loaded_from
+      source.path.join(relative_loaded_from).to_s
+    end
+
+    def full_gem_path
+      Pathname.new(loaded_from).dirname.expand_path.to_s
+    end
+
+  end
+
   module SharedHelpers
     attr_accessor :gem_loaded