+++ spec/runtime/load_spec.rb
@@ -1,7 +1,6 @@
 require File.expand_path('../../spec_helper', __FILE__)
 
 describe "Bundler.load" do
-
   before :each do
     system_gems "rack-1.0.0"
     # clear memoized method results
@@ -52,7 +51,7 @@ describe "Bundler.load" do
         gem "activesupport", :group => :test
       G
 
-      ruby <<-R
+      ruby <<-RUBY
         require "bundler"
         Bundler.setup :default
         Bundler.require :default
@@ -62,20 +61,45 @@ describe "Bundler.load" do
         rescue LoadError
           puts "no activesupport"
         end
-      R
+      RUBY
 
       out.split("\n").should == ["1.0.0", "no activesupport"]
     end
   end
 
-  describe "not hurting brittle rubygems" do
+  describe "when locked" do
     before :each do
-      system_gems "activerecord-2.3.2", "activesupport-2.3.2"
+      install_gemfile <<-G
+        source "file://#{gem_repo1}"
+        gem "activesupport"
+      G
+      bundle :lock
     end
 
+    it "loads env.rb instead of the runtime" do
+      ruby <<-RUBY
+        require 'bundler'
+        Bundler.load
+        puts Bundler.instance_eval{ @runtime }.inspect
+      RUBY
+      out.should == "nil"
+    end
+
+    it "does not invoke setup inside env.rb" do
+      ruby <<-RUBY
+        require 'bundler'
+        Bundler.load
+        puts $LOAD_PATH.grep(/activesupport/i)
+      RUBY
+
+      out.should == ""
+    end
+  end
+
+  describe "not hurting brittle rubygems" do
     it "does not inject #source into the generated YAML of the gem specs" do
+      system_gems "activerecord-2.3.2", "activesupport-2.3.2"
       gemfile <<-G
-        source "file://#{gem_repo1}"
         gem "activerecord"
       G
 
@@ -85,4 +109,5 @@ describe "Bundler.load" do
       end
     end
   end
+
 end